refactor: centralize map test reactor definitions into a shared support file
