+++ spec/single_worker_map_spec.rb
@@ -2,28 +2,6 @@
 
 require "spec_helper"
 
-# Define a simple reactor for mapping
-class SingleWorkerDoubleReactor < RubyReactor::Reactor
-  input :number
-
-  step :double do
-    argument :number, input(:number)
-    run { |args, _| RubyReactor::Success(args[:number] * 2) }
-  end
-
-  returns :double
-end
-
-class SingleWorkerMapReactor < RubyReactor::Reactor
-  input :numbers
-
-  map :doubled_numbers, SingleWorkerDoubleReactor do
-    source input(:numbers)
-    argument :number, element(:doubled_numbers)
-    async true # Default single worker strategy
-  end
-end
-
 RSpec.describe "Single Worker Async Map Execution" do
   before do
     # Ensure clean state
@@ -32,7 +10,7 @@ RSpec.describe "Single Worker Async Map Execution" do
 
   it "queues MapExecutionWorker for single worker strategy" do
     # Run initial execution
-    reactor = SingleWorkerMapReactor.new
+    reactor = MapTestReactors::SingleWorkerMapReactor.new
     reactor.run(numbers: [1, 2, 3])
     context_id = reactor.context.context_id
 
@@ -44,17 +22,17 @@ RSpec.describe "Single Worker Async Map Execution" do
     job = RubyReactor::SidekiqWorkers::MapExecutionWorker.jobs.first
     args = job["args"].first
     expect(args["serialized_inputs"]).to be_a(Hash)
-    expect(args["reactor_class_info"]).to eq({ "type" => "class", "name" => "SingleWorkerDoubleReactor" })
+    expect(args["reactor_class_info"]).to eq({ "type" => "class", "name" => "MapTestReactors::DoubleReactor" })
 
     # Process the job
     RubyReactor::SidekiqWorkers::MapExecutionWorker.drain
 
     # Verify result in Redis
     storage = RubyReactor.configuration.storage_adapter
-    context = storage.retrieve_context(context_id, SingleWorkerMapReactor.name)
+    context = storage.retrieve_context(context_id, MapTestReactors::SingleWorkerMapReactor.name)
 
     expect(context).not_to be_nil
-    # Since SingleWorkerDoubleReactor returns :double, the result should be [2, 4, 6]
+    # Since DoubleReactor returns :double, the result should be [2, 4, 6]
     expect(context["intermediate_results"]["doubled_numbers"]).to eq([2, 4, 6])
   end
 end
